refactor apis
