+++ api/common/utils.js
@@ -34,3 +34,31 @@ export function concatHexStrings(hexStrings) {
   }
   return "0x" + result;
 }
+
+export function currentNpmScriptName() {
+  return process.env.npm_lifecycle_event;
+}
+
+export function logReceiptAndEvents(rawreceiptList, blockid, matchedEventOffsets, filteredEventList) {
+  console.log(
+    "[*]",
+    rawreceiptList.length,
+    rawreceiptList.length > 1
+      ? "receipts fetched from block"
+      : "receipt fetched from block",
+    blockid
+  );
+  console.log(
+    "[*]",
+    matchedEventOffsets.length / 7,
+    matchedEventOffsets.length / 7 > 1 ? "events matched" : "event matched"
+  );
+  for (let i in filteredEventList) {
+    for (let j in filteredEventList[i]) {
+      filteredEventList[i][j].prettyPrint(
+        "\tTx[" + i + "]Event[" + j + "]",
+        false
+      );
+    }
+  }
+}
