+++ package.json
@@ -1,14 +1,14 @@
 {
   "name": "zkgraph",
   "scripts": {
-    "compile-local": "node api/api-local/compile.js",
-    "exec-local": "node api/api-local/exec.js",
-    "prove-local": "node api/api-local/prove.js",
+    "compile-local": "node api/compile.js",
+    "exec-local": "node api/exec.js",
+    "prove-local": "node api/prove.js",
     "deploy-local": "",
-    "compile": "node api/api-full/compile.js",
-    "setup": "node api/api-full/setup.js",
-    "exec": "node api/api-full/exec.js",
-    "prove": "node api/api-full/prove.js",
+    "compile": "node api/compile.js",
+    "setup": "node api/setup.js",
+    "exec": "node api/exec.js",
+    "prove": "node api/prove.js",
     "deploy": "",
     "verify": "",
     "format": "prettier --write -c **/*.{js,ts,md}",
